requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
PyQt5==5.15.10
numpy==1.26.4
opencv-python==4.8.0.74
soundfile==0.12.1
sounddevice==0.4.6
psutil==5.9.6
wmi==1.5.1
python-dotenv==1.0.0
jinja2==3.1.2
pyyaml==6.0.1
scikit-learn==1.3.2
pydub==0.25.1
aiofiles==23.2.1
httpx==0.25.2
sse-starlette==1.8.2
concurrently
